requirements.txt
numpy==1.26.3
Pillow==10.2.0
streamlit==1.30.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
